refactor: AcceptanceTest의 픽스쳐를 각 클래스로 이동
