+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -4,7 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.admin.dto.StationResponse;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static io.restassured.RestAssured.*;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -36,6 +38,17 @@ public class StationAcceptanceTest extends AcceptanceTest {
                         log().all().
                         extract().
                         jsonPath().getList(".", StationResponse.class);
+    }
+
+    StationResponse createStation(String name) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+
+        return post(
+                "/stations",
+                params,
+                StationResponse.class
+        );
     }
 
     void deleteStation(Long id) {
